refactor(agent): extract transport mode dispatch module (#228)
